Isolate SQLite storage boundaries
