+++ src/mindroom/agent_storage.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING, Any, cast
 
-from agno.db.base import SessionType
+from agno.db.base import BaseDb, SessionType
 from agno.db.sqlite import SqliteDb
 from agno.learn import LearningMachine
 from agno.session.agent import AgentSession
@@ -22,23 +22,40 @@ if TYPE_CHECKING:
     from mindroom.tool_system.worker_routing import ToolExecutionIdentity
 
 __all__ = [
+    "create_culture_storage",
     "create_session_storage",
-    "create_state_storage_db",
-    "get_agent_runtime_sqlite_dbs",
+    "create_state_storage",
+    "get_agent_runtime_state_dbs",
     "get_agent_session",
     "get_team_session",
 ]
 
 
-def get_agent_runtime_sqlite_dbs(agent: Agent) -> tuple[SqliteDb | None, SqliteDb | None]:
-    """Return the runtime-owned SQLite handles attached to one agent."""
-    history_db = agent.db if isinstance(agent.db, SqliteDb) else None
+def get_agent_runtime_state_dbs(agent: Agent) -> tuple[BaseDb | None, BaseDb | None]:
+    """Return the runtime-owned Agno DB handles attached to one agent."""
+    history_db = agent.db if isinstance(agent.db, BaseDb) else None
     learning = agent.learning
-    learning_db = learning.db if isinstance(learning, LearningMachine) and isinstance(learning.db, SqliteDb) else None
+    learning_db = learning.db if isinstance(learning, LearningMachine) and isinstance(learning.db, BaseDb) else None
     return history_db, learning_db
 
 
-def create_state_storage_db(
+def create_state_storage(
+    storage_name: str,
+    state_root: Path,
+    *,
+    subdir: str,
+    session_table: str,
+) -> BaseDb:
+    """Create persistent Agno state storage from an already-resolved state root."""
+    return _create_sqlite_state_storage(
+        storage_name=storage_name,
+        state_root=state_root,
+        subdir=subdir,
+        session_table=session_table,
+    )
+
+
+def _create_sqlite_state_storage(
     storage_name: str,
     state_root: Path,
     *,
@@ -56,7 +73,7 @@ def create_session_storage(
     config: Config,
     runtime_paths: RuntimePaths,
     execution_identity: ToolExecutionIdentity | None,
-) -> SqliteDb:
+) -> BaseDb:
     """Create persistent session storage for an agent."""
     return _create_agent_state_db(
         agent_name,
@@ -76,15 +93,15 @@ def _create_agent_state_db(
     *,
     subdir: str,
     session_table: str,
-) -> SqliteDb:
-    """Create a persistent SQLite database for one agent state category."""
+) -> BaseDb:
+    """Create persistent storage for one agent state category."""
     state_storage_path = resolve_agent_runtime(
         agent_name,
         config,
         runtime_paths,
         execution_identity=execution_identity,
     ).state_root
-    return create_state_storage_db(
+    return create_state_storage(
         storage_name=agent_name,
         state_root=state_storage_path,
         subdir=subdir,
@@ -92,7 +109,14 @@ def _create_agent_state_db(
     )
 
 
-def get_agent_session(storage: SqliteDb, session_id: str) -> AgentSession | None:
+def create_culture_storage(culture_name: str, storage_path: Path) -> BaseDb:
+    """Create persistent culture storage shared by all agents in a culture."""
+    culture_dir = storage_path / "culture"
+    culture_dir.mkdir(parents=True, exist_ok=True)
+    return SqliteDb(db_file=str(culture_dir / f"{culture_name}.db"))
+
+
+def get_agent_session(storage: BaseDb, session_id: str) -> AgentSession | None:
     """Retrieve and deserialize an AgentSession from storage."""
     raw = storage.get_session(session_id, SessionType.AGENT)
     if raw is None:
@@ -104,7 +128,7 @@ def get_agent_session(storage: SqliteDb, session_id: str) -> AgentSession | None
     return None
 
 
-def get_team_session(storage: SqliteDb, session_id: str) -> TeamSession | None:
+def get_team_session(storage: BaseDb, session_id: str) -> TeamSession | None:
     """Retrieve and deserialize a TeamSession from storage."""
     raw = storage.get_session(session_id, SessionType.TEAM)
     if raw is None:
