+++ src/mindroom/history/__init__.py
@@ -6,8 +6,8 @@ from mindroom.history.runtime import (
     PreparedScopeHistory,
     ScopeSessionContext,
     apply_replay_plan,
-    close_agent_runtime_sqlite_dbs,
-    close_team_runtime_sqlite_dbs,
+    close_agent_runtime_state_dbs,
+    close_team_runtime_state_dbs,
     create_scope_session_storage,
     estimate_preparation_static_tokens,
     estimate_preparation_static_tokens_for_team,
@@ -64,8 +64,8 @@ __all__ = [
     "ScopeSessionContext",
     "add_pending_force_compaction_scope",
     "apply_replay_plan",
-    "close_agent_runtime_sqlite_dbs",
-    "close_team_runtime_sqlite_dbs",
+    "close_agent_runtime_state_dbs",
+    "close_team_runtime_state_dbs",
     "compute_prompt_token_breakdown",
     "create_scope_session_storage",
     "estimate_preparation_static_tokens",
