+++ tests/test_agents.py
@@ -12,11 +12,12 @@ from unittest.mock import MagicMock, patch
 
 import pytest
 from agno.agent import Agent
+from agno.db.in_memory import InMemoryDb
 from agno.learn import LearningMachine, LearningMode, UserMemoryConfig, UserProfileConfig
 from pydantic import ValidationError
 
 from mindroom import agent_prompts
-from mindroom.agent_storage import get_agent_runtime_sqlite_dbs
+from mindroom.agent_storage import get_agent_runtime_state_dbs
 from mindroom.agents import (
     _CULTURE_MANAGER_CACHE,
     _PRIVATE_CULTURE_MANAGER_CACHE,
@@ -204,7 +205,7 @@ def test_config_round_trips_structured_agent_tool_entries() -> None:
     }
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_get_agent_calculator(mock_storage: MagicMock) -> None:  # noqa: ARG001
     """Tests that the calculator agent is created correctly."""
     config = _test_config()
@@ -213,7 +214,7 @@ def test_get_agent_calculator(mock_storage: MagicMock) -> None:  # noqa: ARG001
     assert agent.name == "CalculatorAgent"
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_get_agent_general(mock_storage: MagicMock) -> None:  # noqa: ARG001
     """Tests that the general agent is created correctly."""
     config = _test_config()
@@ -227,7 +228,16 @@ def test_get_agent_general(mock_storage: MagicMock) -> None:  # noqa: ARG001
     assert agent.learning.user_memory.mode is LearningMode.ALWAYS
 
 
-def test_get_agent_runtime_sqlite_dbs_includes_learning_storage(tmp_path: Path) -> None:
+def test_get_agent_runtime_state_dbs_accepts_backend_neutral_agno_storage() -> None:
+    """Runtime DB discovery should not require SQLite-specific storage handles."""
+    history_db = InMemoryDb()
+    learning_db = InMemoryDb()
+    agent = Agent(db=history_db, learning=LearningMachine(db=learning_db))
+
+    assert get_agent_runtime_state_dbs(agent) == (history_db, learning_db)
+
+
+def test_get_agent_runtime_state_dbs_includes_learning_storage(tmp_path: Path) -> None:
     """Runtime-owned agent DB cleanup should include the separate learning storage."""
     config = _test_config()
     config.models = {"default": ModelConfig(provider="ollama", id="test-model")}
@@ -236,7 +246,7 @@ def test_get_agent_runtime_sqlite_dbs_includes_learning_storage(tmp_path: Path) 
     agent = _create_agent_for_test("general", config=config)
 
     assert isinstance(agent.learning, LearningMachine)
-    history_db, learning_db = get_agent_runtime_sqlite_dbs(agent)
+    history_db, learning_db = get_agent_runtime_state_dbs(agent)
     assert history_db is agent.db
     assert learning_db is agent.learning.db
     assert history_db is not learning_db
@@ -247,7 +257,7 @@ def test_get_agent_runtime_sqlite_dbs_includes_learning_storage(tmp_path: Path) 
         history_db.close()
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_hidden_tool_calls_prompt_is_injected(mock_storage: MagicMock) -> None:  # noqa: ARG001
     """Agents with hidden tool calls get a prompt hint to avoid narrating tool usage."""
     config = _test_config()
@@ -291,7 +301,7 @@ def test_history_limits_require_positive_values(factory: Callable[[], object], f
         factory()
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_scheduler_tool_enabled_by_default(mock_storage: MagicMock) -> None:  # noqa: ARG001
     """All agents should get the scheduler tool even when not explicitly configured."""
     config = _test_config()
@@ -303,7 +313,7 @@ def test_scheduler_tool_enabled_by_default(mock_storage: MagicMock) -> None:  # 
     assert "scheduler" in tool_names
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_configurable_default_tools_are_applied(mock_storage: MagicMock) -> None:  # noqa: ARG001
     """defaults.tools should be merged into every agent's configured tools."""
     config = _test_config()
@@ -317,7 +327,7 @@ def test_configurable_default_tools_are_applied(mock_storage: MagicMock) -> None
     assert "calculator" in tool_names
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_default_tools_do_not_duplicate_agent_tools(mock_storage: MagicMock) -> None:  # noqa: ARG001
     """An agent tool already present should not be duplicated by defaults.tools."""
     config = _test_config()
@@ -330,7 +340,7 @@ def test_default_tools_do_not_duplicate_agent_tools(mock_storage: MagicMock) -> 
     assert tool_names.count("scheduler") == 1
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_agent_include_default_tools_false_skips_config_defaults(mock_storage: MagicMock) -> None:  # noqa: ARG001
     """Agent include_default_tools=False should skip defaults.tools entirely."""
     config = _test_config()
@@ -394,7 +404,7 @@ def test_openclaw_compat_expansion_dedupes_preserving_order() -> None:
 
 
 @patch("mindroom.agents.get_tool_by_name")
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_uses_native_tool_lookups_for_openclaw_compat(
     mock_storage: MagicMock,  # noqa: ARG001
     mock_get_tool_by_name: MagicMock,
@@ -412,7 +422,7 @@ def test_create_agent_uses_native_tool_lookups_for_openclaw_compat(
 
 
 @patch("mindroom.agents.get_tool_by_name")
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_passes_merged_tool_config_overrides_to_registered_tools(
     mock_storage: MagicMock,  # noqa: ARG001
     mock_get_tool_by_name: MagicMock,
@@ -435,7 +445,7 @@ def test_create_agent_passes_merged_tool_config_overrides_to_registered_tools(
 
 
 @patch("mindroom.agents.get_tool_by_name")
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_keeps_runtime_base_dir_separate_from_authored_tool_config(
     mock_storage: MagicMock,  # noqa: ARG001
     mock_get_tool_by_name: MagicMock,
@@ -461,7 +471,7 @@ def test_create_agent_keeps_runtime_base_dir_separate_from_authored_tool_config(
 
 
 @patch("mindroom.agents.get_tool_by_name")
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_continues_when_implied_tool_import_fails(
     mock_storage: MagicMock,  # noqa: ARG001
     mock_get_tool_by_name: MagicMock,
@@ -517,7 +527,7 @@ def test_create_agent_continues_when_implied_tool_import_fails(
 
 
 @patch("mindroom.agents.get_tool_by_name")
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_continues_when_tool_lookup_reports_unknown_tool(
     mock_storage: MagicMock,  # noqa: ARG001
     mock_get_tool_by_name: MagicMock,
@@ -569,7 +579,7 @@ def test_create_agent_continues_when_tool_lookup_reports_unknown_tool(
 
 
 @patch("mindroom.agents.get_tool_by_name")
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_expands_openclaw_compat_for_worker_tool_overrides(
     mock_storage: MagicMock,  # noqa: ARG001
     mock_get_tool_by_name: MagicMock,
@@ -590,7 +600,7 @@ def test_create_agent_expands_openclaw_compat_for_worker_tool_overrides(
 
 
 @patch("mindroom.agents.get_tool_by_name")
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_uses_memory_file_workspace_for_base_dir_tools(
     mock_storage: MagicMock,  # noqa: ARG001
     mock_get_tool_by_name: MagicMock,
@@ -1290,7 +1300,7 @@ def test_private_workspace_template_initializes_missing_files_in_partially_popul
 
 
 @patch("mindroom.agents.get_tool_by_name")
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_does_not_pass_browser_specific_runtime_overrides(
     mock_storage: MagicMock,  # noqa: ARG001
     mock_get_tool_by_name: MagicMock,
@@ -1313,7 +1323,7 @@ def test_create_agent_does_not_pass_browser_specific_runtime_overrides(
 
 
 @patch("mindroom.agents.get_tool_by_name")
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_passes_authored_shell_runtime_overrides(
     mock_storage: MagicMock,  # noqa: ARG001
     mock_get_tool_by_name: MagicMock,
@@ -1344,7 +1354,7 @@ def test_create_agent_passes_authored_shell_runtime_overrides(
 
 
 @patch("mindroom.agents.get_tool_by_name")
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_keeps_tool_default_base_dir_without_memory_workspace(
     mock_storage: MagicMock,  # noqa: ARG001
     mock_get_tool_by_name: MagicMock,
@@ -1377,7 +1387,7 @@ def test_create_agent_threads_config_path_to_plugin_loading(
     config = _test_config()
     runtime_paths = _runtime_paths(tmp_path, config_path=config_path)
 
-    with patch("mindroom.agents.SqliteDb"):
+    with patch("mindroom.agent_storage.SqliteDb"):
         _create_agent_for_test("general", config=_bind_runtime_paths(config, runtime_paths))
 
     mock_load_plugins.assert_called_once()
@@ -1424,7 +1434,7 @@ def test_create_agent_rejects_bare_env_var_context_files() -> None:
 
 
 @patch("mindroom.agents.get_tool_by_name")
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_applies_agent_workspace_override_for_worker_routed_scoped_tools(
     mock_storage: MagicMock,  # noqa: ARG001
     mock_get_tool_by_name: MagicMock,
@@ -1451,7 +1461,7 @@ def test_create_agent_applies_agent_workspace_override_for_worker_routed_scoped_
 
 
 @patch("mindroom.agents.get_tool_by_name")
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_uses_default_worker_tool_policy_when_unset(
     mock_storage: MagicMock,  # noqa: ARG001
     mock_get_tool_by_name: MagicMock,
@@ -1470,7 +1480,7 @@ def test_create_agent_uses_default_worker_tool_policy_when_unset(
     assert all(override == ["shell", "coding"] for override in worker_overrides)
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_openclaw_compat_implies_matrix_message_tool(mock_storage: MagicMock) -> None:  # noqa: ARG001
     """openclaw_compat should stay in the runtime toolkit list and imply matrix_message."""
     config = _test_config()
@@ -1516,7 +1526,7 @@ def test_matrix_message_implied_attachments_does_not_duplicate() -> None:
     assert effective_tools.count("attachments") == 1
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_get_agent_code(mock_storage: MagicMock) -> None:  # noqa: ARG001
     """Tests that the code agent is created correctly."""
     config = _test_config()
@@ -1525,7 +1535,7 @@ def test_get_agent_code(mock_storage: MagicMock) -> None:  # noqa: ARG001
     assert agent.name == "CodeAgent"
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_get_agent_shell(mock_storage: MagicMock) -> None:  # noqa: ARG001
     """Tests that the shell agent is created correctly."""
     config = _test_config()
@@ -1534,7 +1544,7 @@ def test_get_agent_shell(mock_storage: MagicMock) -> None:  # noqa: ARG001
     assert agent.name == "ShellAgent"
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_get_agent_summary(mock_storage: MagicMock) -> None:  # noqa: ARG001
     """Tests that the summary agent is created correctly."""
     config = _test_config()
@@ -1576,7 +1586,7 @@ def test_get_agent_learning_defaults_fallback_when_agent_setting_omitted(mock_st
     assert mock_storage.call_count == 1
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_get_agent_learning_agentic_mode(mock_storage: MagicMock) -> None:  # noqa: ARG001
     """Tests that learning mode can be configured as agentic."""
     config = _test_config()
@@ -1682,7 +1692,7 @@ def test_get_agent_uses_shared_worker_storage_without_execution_identity(
     assert agent_root / "learning" / "general.db" in db_files
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_loads_shared_worker_scoped_tool_credentials_with_explicit_shared_identity(
     mock_storage: MagicMock,  # noqa: ARG001
     tmp_path: Path,
@@ -1907,7 +1917,7 @@ def test_resolve_agent_owned_path_rejects_path_traversal(tmp_path: Path) -> None
 
 
 @patch("mindroom.agents.get_tool_by_name")
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_reads_canonical_context_files_and_reloads_from_agent_root(
     mock_storage: MagicMock,  # noqa: ARG001
     mock_get_tool_by_name: MagicMock,
@@ -1963,7 +1973,7 @@ def test_create_agent_reads_canonical_context_files_and_reloads_from_agent_root(
     assert "Updated canonical soul context." not in deleted_agent.role
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_scaffolds_default_mind_workspace_under_runtime_storage_root(
     _mock_storage: MagicMock,  # noqa: PT019
     tmp_path: Path,
@@ -2085,7 +2095,7 @@ def test_create_agent_uses_mounted_dedicated_worker_root_for_unscoped_agent_stat
     assert mock_get_tool_by_name.call_args.kwargs["tool_init_overrides"] == {"base_dir": str(canonical_workspace)}
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_agent_context_files_are_loaded_into_role(mock_storage: MagicMock, tmp_path: Path) -> None:  # noqa: ARG001
     """Context files should load directly from the canonical workspace."""
     config = _test_config()
@@ -2112,7 +2122,7 @@ def test_agent_context_files_are_loaded_into_role(mock_storage: MagicMock, tmp_p
     assert "Canonical soul directive." in updated_agent.role
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_agent_preload_cap_truncates_context_files_in_order(
     mock_storage: MagicMock,  # noqa: ARG001
     tmp_path: Path,
@@ -2140,7 +2150,7 @@ def test_agent_preload_cap_truncates_context_files_in_order(
     assert "SECOND_START" in agent.role
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_agent_missing_context_file_is_ignored(mock_storage: MagicMock, tmp_path: Path) -> None:  # noqa: ARG001
     """Missing context files should not prevent agent creation."""
     config = _test_config()
@@ -2167,7 +2177,7 @@ def test_agent_relative_context_paths_resolve_from_workspace_not_cwd(tmp_path: P
     other_cwd.mkdir(parents=True, exist_ok=True)
     os.chdir(other_cwd)
     try:
-        with patch("mindroom.agents.SqliteDb"):
+        with patch("mindroom.agent_storage.SqliteDb"):
             agent = _create_agent_for_test("general", config=_bind_runtime_paths(config, _runtime_paths(tmp_path)))
     finally:
         os.chdir(original_cwd)
@@ -2299,7 +2309,7 @@ def test_copy_workspace_template_rejects_destination_symlink_escape(tmp_path: Pa
         _copy_workspace_template(workspace_root, template_dir=template_dir)
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_private_root_loads_requester_context_from_isolated_workspace(
     mock_storage: MagicMock,  # noqa: ARG001
     tmp_path: Path,
@@ -2398,7 +2408,7 @@ def test_create_agent_private_root_loads_requester_context_from_isolated_workspa
     assert alice_workspace.parent == private_instance_scope_root_path(tmp_path, alice_worker_key) / "general"
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_private_template_dir_does_not_imply_context_files(
     mock_storage: MagicMock,  # noqa: ARG001
     tmp_path: Path,
@@ -2445,7 +2455,7 @@ def test_create_agent_private_template_dir_does_not_imply_context_files(
     assert "Template user." not in agent.role
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 def test_create_agent_private_root_requires_execution_identity(
     mock_storage: MagicMock,  # noqa: ARG001
     tmp_path: Path,
@@ -3228,7 +3238,7 @@ def test_config_rejects_git_backed_private_knowledge_overlapping_template_conten
         )
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 @patch("mindroom.agents.CultureManager")
 @patch("mindroom.agents.Agent")
 def test_create_agent_shares_culture_manager_for_same_culture(
@@ -3300,7 +3310,7 @@ def test_create_agent_shares_culture_manager_for_same_culture(
     assert len(culture_db_calls) == 1
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 @patch("mindroom.agents.CultureManager")
 @patch("mindroom.agents.Agent")
 def test_create_agent_culture_uses_agent_model_when_default_missing(
@@ -3347,12 +3357,14 @@ def test_create_agent_culture_uses_agent_model_when_default_missing(
     call_args = mock_get_model_instance.call_args
     assert call_args.args[2] == "m1"  # model_name
     assert mock_agent_class.call_count == 1
-    assert mock_storage.call_count == 1
+    db_files = [Path(str(call.kwargs["db_file"])) for call in mock_storage.call_args_list]
+    assert agent_state_root_path(tmp_path, "agent_one") / "sessions" / "agent_one.db" in db_files
+    assert tmp_path / "culture" / "engineering.db" in db_files
     assert mock_culture_manager_class.call_args is not None
     assert mock_culture_manager_class.call_args.kwargs["model"] is model
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 @patch("mindroom.agents.CultureManager")
 @patch("mindroom.agents.Agent")
 def test_create_private_agent_scopes_culture_storage_per_requester(
@@ -3443,7 +3455,7 @@ def test_create_private_agent_scopes_culture_storage_per_requester(
     assert second_kwargs["culture_manager"] is created_culture_managers[1]
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 @patch("mindroom.agents.CultureManager")
 @patch("mindroom.agents.Agent")
 def test_private_agents_share_culture_manager_within_same_requester_scope(
@@ -3538,7 +3550,7 @@ def test_private_agents_share_culture_manager_within_same_requester_scope(
     assert second_kwargs["culture_manager"] is created_culture_manager
 
 
-@patch("mindroom.agents.SqliteDb")
+@patch("mindroom.agent_storage.SqliteDb")
 @patch("mindroom.agents.CultureManager")
 @patch("mindroom.agents.Agent")
 def test_private_user_agent_agents_share_culture_manager_within_same_requester_scope(
