+++ tests/test_queued_message_notify.py
@@ -1097,7 +1097,7 @@ async def test_ai_response_preserves_stale_notice_before_prepare(tmp_path: Path)
             side_effect=lambda **_kwargs: _open_scope(storage),
         ),
         patch("mindroom.ai._prepare_agent_and_prompt", new=AsyncMock(side_effect=fake_prepare)),
-        patch("mindroom.ai.close_agent_runtime_sqlite_dbs"),
+        patch("mindroom.ai.close_agent_runtime_state_dbs"),
     ):
         response = await ai_response(
             agent_name="general",
@@ -1161,7 +1161,7 @@ async def test_ai_response_preserves_notice_in_run_output_and_session(tmp_path: 
             side_effect=lambda **_kwargs: _open_scope(storage),
         ),
         patch("mindroom.ai._prepare_agent_and_prompt", new=AsyncMock(return_value=_prepared_run(agent))),
-        patch("mindroom.ai.close_agent_runtime_sqlite_dbs"),
+        patch("mindroom.ai.close_agent_runtime_state_dbs"),
         queued_message_signal_context(_StaticQueuedState(pending=True)),
     ):
         response = await ai_response(
@@ -1214,7 +1214,7 @@ async def test_ai_response_preserves_notice_in_session_after_exception(tmp_path:
             side_effect=lambda **_kwargs: _open_scope(storage),
         ),
         patch("mindroom.ai._prepare_agent_and_prompt", new=AsyncMock(return_value=_prepared_run(agent))),
-        patch("mindroom.ai.close_agent_runtime_sqlite_dbs"),
+        patch("mindroom.ai.close_agent_runtime_state_dbs"),
         queued_message_signal_context(_StaticQueuedState(pending=True)),
     ):
         response = await ai_response(
@@ -1267,7 +1267,7 @@ async def test_stream_agent_response_preserves_notice_in_session(tmp_path: Path)
             side_effect=lambda **_kwargs: _open_scope(storage),
         ),
         patch("mindroom.ai._prepare_agent_and_prompt", new=AsyncMock(return_value=_prepared_run(agent))),
-        patch("mindroom.ai.close_agent_runtime_sqlite_dbs"),
+        patch("mindroom.ai.close_agent_runtime_state_dbs"),
         queued_message_signal_context(_StaticQueuedState(pending=True)),
     ):
         chunks = [
