+++ tests/test_tool_dependencies.py
@@ -123,7 +123,7 @@ def test_core_runtime_imports_are_declared_as_base_dependencies() -> None:
     required_runtime_dependencies = {
         "src/mindroom/constants.py": "pydantic-settings",
         "src/mindroom/matrix/client.py": "aiohttp",
-        "src/mindroom/matrix/_event_cache.py": "aiosqlite",
+        "src/mindroom/matrix/cache/sqlite_event_cache.py": "aiosqlite",
         "src/mindroom/mcp/transports.py": "anyio",
     }
